Fixed annoying Vundle untracked changes
It should be fixed in Vundle itself. It now sets the default path to
bundle/vundle somehow, so the spell files are saved there. By allowing a
dirty submodule it won't bother me when committing my dotfiles, but I
will be losing my spell custom dicts :(

A better fix would be to set the spellfile manually in my vimrc.